+++ src/main/java/ddd/caffeine/ratrip/module/auth/external/KakaoUserApiClient.java
@@ -4,7 +4,7 @@ import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
 
-import ddd.caffeine.ratrip.module.auth.external.dto.KakaoProfileResponse;
+import ddd.caffeine.ratrip.module.auth.external.dto.response.KakaoProfileResponse;
 
 @FeignClient(name = "kakaoUserApiClient", url = "https://kapi.kakao.com")
 public interface KakaoUserApiClient {
