+++ src/main/java/ddd/caffeine/ratrip/module/auth/external/dto/response/KakaoAccount.java
@@ -1,0 +1,9 @@
+package ddd.caffeine.ratrip.module.auth.external.dto.response;
+
+import lombok.Getter;
+
+@Getter
+public class KakaoAccount {
+	private Profile profile;
+	private String email;
+}
