+++ src/main/java/ddd/caffeine/ratrip/module/auth/external/dto/response/KakaoProfileResponse.java
@@ -1,0 +1,13 @@
+package ddd.caffeine.ratrip.module.auth.external.dto.response;
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+
+import lombok.Getter;
+
+@Getter
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+public class KakaoProfileResponse {
+	private String id;
+	private KakaoAccount kakaoAccount;
+}
