+++ src/main/java/ddd/caffeine/ratrip/module/auth/external/dto/response/Profile.java
@@ -1,0 +1,9 @@
+package ddd.caffeine.ratrip.module.auth.external.dto.response;
+
+import lombok.Getter;
+
+@Getter
+public class Profile {
+	private String nickname;
+}
+
